pyproject.toml
[project]
name = "djangotesttask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "channels[daphne]>=4.1.0",
    "django-cors-headers>=4.5.0",
    "django==4.2",
    "djangorestframework>=3.15.2",
    "drf-yasg>=1.21.8",
    "pip>=24.2",
]
